backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
torch==2.9.1
torchvision==0.24.1

# Image Processing
Pillow==11.3.0
opencv-python==4.8.1.78
numpy==1.24.3

# Validation
pydantic==2.5.0
python-dotenv==1.0.0

